+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 ### Added
 - Add `Shaped::Array` method and class
+- Add `Array` refinement
 
 ## 0.1.0 - 2020-06-16
 - Initial release of Shaped! Validate the shape of Ruby hashes and/or arrays.