+++ CHANGELOG.md
@@ -4,5 +4,8 @@
 - Add `Array` refinement
 - Add `Shaped::MatchFailureReason` and `Shaped::Array#match_failure_reason`
 
+### Fixed
+- Fix potential memory leak in long-lived `Shaped::Array` instances
+
 ## 0.1.0 - 2020-06-16
 - Initial release of Shaped! Validate the shape of Ruby hashes and/or arrays.